Create getOS utility function
